refactor: split recovery cache policy
