+++ .agents/scripts/tests/test-profile-readme-boundary.sh
@@ -94,7 +94,9 @@ install_helper_with_libs() {
 		"${SOURCE_HELPER%/*}/screen_time_linux_logind.py" \
 		"${SOURCE_HELPER%/*}/screen_time_linux_wtmp.py" \
 		"${SOURCE_HELPER%/*}/screen_time_history.py" \
-		"${SOURCE_HELPER%/*}/worktree-recovery-cache-policy.py" "$helper_dir/"
+		"${SOURCE_HELPER%/*}/worktree-recovery-cache-policy.py" \
+		"${SOURCE_HELPER%/*}/worktree_recovery_cache_policy_common.py" \
+		"${SOURCE_HELPER%/*}/worktree_recovery_cache_policy_operations.py" "$helper_dir/"
 	# Keep this profile-boundary fixture independent of host /proc visibility.
 	# Shared worktree-removal guard behaviour has dedicated tests; this stub proves
 	# the profile helper invokes cleanup from the canonical checkout after its
